+++ IceAndFire/app/__factory__/index.js
@@ -1,0 +1,3 @@
+/*
+Will contain static factory data for test
+*/
