+++ IceAndFire/app/components/Main.js
@@ -1,0 +1,46 @@
+import {
+    Platform,
+    StyleSheet,
+    Text,
+    View
+} from 'react-native';
+
+import React, { Component } from 'react';
+
+export default class Main extends Component {
+    render() {
+        return (
+            <View style={styles.container}>
+                <Text style={styles.welcome}>
+                    Welcome to React Native!
+            </Text>
+                <Text style={styles.instructions}>
+                    To get started, edit App.js
+            </Text>
+                <Text style={styles.instructions}>
+                    {instructions}
+                </Text>
+            </View>
+        );
+    }
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+        backgroundColor: '#F5FCFF',
+    },
+    welcome: {
+        fontSize: 20,
+        textAlign: 'center',
+        margin: 10,
+    },
+    instructions: {
+        textAlign: 'center',
+        color: '#333333',
+        marginBottom: 5,
+    },
+});
+
